pyproject.toml
[tool.poetry]
name = "learning-phy-distances"
version = "0.1.0"
description = ""
authors = ["Benjamin Bjoern Kroeger <benjamin_kroeger@t-online.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.2"
torchvision = "^0.17.2"
torchaudio = "^2.2.2"
pandas = "^2.2.2"
numpy = "^1.26.4"
scipy = "^1.13.0"
scikit-learn = "^1.4.2"
ete3 = "^3.1.3" # ete4
transformers = "^4.40.1"
matplotlib = "^3.8.4"
fair-esm = "^2.0.0"
treeswift = "^1.1.42"
pyqt5-qt5 = "=5.15.2"
pyqt5 = "=5.15.10"
ete4 = {git = "https://github.com/etetoolkit/ete.git",branch="ete4"}

# itol
pyqt6 = "^6.6.1"
biopython = "^1.83"
umap = "^0.1.1"
datashader = "^0.16.1"
bokeh = "^3.4.1"
holoviews = "^1.18.3"
scikit-image = "^0.23.2"
dask = {extras = ["dataframe"], version = "^2024.4.2"}
h5py = "^3.11.0"
sentencepiece = "^0.2.0"
lightning = "^2.2.3"
seaborn = "^0.13.2"
tensorboard = "^2.16.2"
wandb = "^0.16.6"



[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipywidgets = "^8.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
